Split programs from helpers
