--- src/schema.json
@@ -1,40 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "description": "",
-  "type": "object",
-  "properties": {
-    "name": {
-      "type": "object",
-      "properties": {
-        "first": {
-          "type": "string",
-          "minLength": 1
-        },
-        "middle": {
-          "type": "string",
-          "minLength": 1
-        },
-        "last": {
-          "type": "string",
-          "minLength": 1
-        }
-      },
-      "required": [
-        "first",
-        "middle",
-        "last"
-      ]
-    },
-    "emails": {
-      "type": "array",
-      "items": {
-        "required": [],
-        "properties": {}
-      }
-    }
-  },
-  "required": [
-    "name",
-    "emails"
-  ]
-}
